modules created for seperate classses. TableService Class for table functions
