+++ index.html
@@ -23,7 +23,7 @@
       integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
       crossorigin="anonymous"
     ></script>
-    <script src="/src/js/userController.js"></script>
+    <script type="module" src="/src/js/userController.js"></script>
     <link rel="stylesheet" href="/src/css/user.css" />
   </head>
   <body>
