+++ package.json
@@ -10,6 +10,7 @@
   },
   "author": "Asfaq Moideen",
   "license": "ISC",
+  "type": "module",
   "dependencies": {
     "highcharts": "^11.4.8",
     "vite": "^5.4.11"
